Emitter refactor
